adds secret-engine-path adapter
